Moved old AKS competetion testing code to new folder
