+++ COMP_2024_TESTING/LR_Colin_TESTING.py
