+++ COMP_2024_TESTING/detections_log.txt
