+++ COMP_2024_TESTING/turtle_log.txt
